requirement.txt
django-2.2.7
pytz-2019.3
sqlparse-0.3.0
pillow-6.2.1
djangorestframework-3.10.3
django-filter-2.2.0
markdown-3.1.1
